+++ app-backend/src/__tests__/user-data-map/resolver.test.ts
@@ -35,6 +35,30 @@ describe('UserDataMap resolver tests.', () => {
         });
     });
 
+    describe('tests for get', () => {
+
+        it('should pass when valid params are passed', async () => {
+            let response: IUserDataMapAttributes = await userDataMapResolver.get({
+                user_id: 'test', 
+                vendor_data_type: 'test'
+            });
+            
+            expect(response.user_id).toEqual('test');
+            expect(response.vendor_data_type).toEqual('test');
+        });
+
+        it('should throw error invalid params are passed', async () => {
+            try {
+                let response: IUserDataMapAttributes = await userDataMapResolver.get({
+                    user_id: 'invalid', 
+                    vendor_data_type: 'test'
+                });
+            } catch (error) {
+                expect(error.message).toEqual('No such user record found');
+            }
+        });
+    });
+
     describe('tests for findOrCreate', () => {
 
         it('should pass when all data is passed', async () => {